+++ character/tests/forms/test_skills.py
@@ -1,7 +1,7 @@
 import pytest
 from django import forms
 
-from character.constants.klasses import Klass
+from character.models.klasses import Klass
 from character.forms.skills import SkillsSelectForm